Refactor data entry into smaller components
